Rename files and add restart function
